requirements.txt
# Core AI/ML frameworks
torch>=2.6.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data processing and analysis
yfinance>=0.2.28
requests>=2.31.0
python-dateutil>=2.8.2

# API and web framework
flask>=3.0.0
flask-cors>=4.0.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization (optional, for analysis)
matplotlib>=3.7.0
seaborn>=0.12.0

# Model monitoring and logging
tensorboard>=2.14.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
